Calculate MSE of KF and Sensor.
By using the true value for the paths of the targets, the user can now determine the MSE between them and the sensor- /KF predictions. In order to obtain the sensor measurements for each target the list "labeled" from the function "scan_with_filter" is returned along with the predictions.